some comments

+++ app.R
@@ -4,6 +4,7 @@ if (!require(parallel))
 if (!require("future"))
   install.packages(future)
 
+# create cluster
 cl <- parallel::makeCluster(1L)
 plan(cluster, workers = cl)
 
@@ -13,19 +14,21 @@ server <- function(input, output) {
     iteration <- as.numeric(input$iteration)
     path <- input$path
     
-    print(iteration)
-    print(path)
-    
-    future::future({
+    future({
+      # simple function to write to file
       writeLog <- function(n, path) {
-        file.remove(path)
+        
+        if (exists(path))
+          file.remove(path)
+        
+        # some example value to write to file
         for (i in 1:n) {
           cat("#", i, "-",  as.character(Sys.time()), "\n", file = path, append = TRUE)
           Sys.sleep(1)
         }
       }
       writeLog(iteration, path)
-    }, globals = c("iteration", "path"))
+    }, globals = c("iteration", "path")) # we want to export those variables to future environment
   })
 }
 ui <- fluidPage(
